refactor: remove react query wrapper
