+++ src/app/common/hooks/account/use-refresh-all-account-data.ts
@@ -1,17 +1,16 @@
 import { useCallback } from 'react';
 import { delay } from '@app/common/utils';
-import { useRefreshAccountData } from '@app/store/accounts/account.hooks';
 import { useCurrentAccountMempool } from '@app/query/mempool/mempool.hooks';
 
+// TODO: Can this be removed? It seems like we should be able
+// to use react-query itself to do this if needed?
 export function useRefreshAllAccountData() {
-  const update = useRefreshAccountData();
   const { refetch } = useCurrentAccountMempool();
   return useCallback(
     async (ms?: number) => {
       if (typeof ms === 'number') await delay(ms);
-      update(null);
       void refetch();
     },
-    [update, refetch]
+    [refetch]
   );
 }
