+++ src/app/pages/fund/fund.tsx
@@ -2,10 +2,8 @@ import { Navigate, Outlet, useLocation, useNavigate } from 'react-router-dom';
 
 import { useRouteHeader } from '@app/common/hooks/use-route-header';
 import { Header } from '@app/components/header';
-import {
-  useCurrentAccount,
-  useCurrentAccountAvailableStxBalance,
-} from '@app/store/accounts/account.hooks';
+import { useCurrentAccount } from '@app/store/accounts/account.hooks';
+import { useCurrentAccountAvailableStxBalance } from '@app/query/balance/balance.hooks';
 import { RouteUrls } from '@shared/route-urls';
 
 import { FundLayout } from './fund.layout';
