+++ src/app/pages/send-tokens/components/asset-search/asset-search.tsx
@@ -4,7 +4,7 @@ import { Box, color, Stack } from '@stacks/ui';
 import { AssetWithMeta } from '@app/common/asset-types';
 import { useTransferableAssets } from '@app/store/assets/asset.hooks';
 import { useSelectedAsset } from '@app/pages/send-tokens/hooks/use-selected-asset';
-import { useCurrentAccountAvailableStxBalance } from '@app/store/accounts/account.hooks';
+import { useCurrentAccountAvailableStxBalance } from '@app/query/balance/balance.hooks';
 
 import { AssetSearchField } from './asset-search-field';
 import { SelectedAsset } from './selected-asset';
