+++ src/app/pages/send-tokens/hooks/use-send-form.ts
@@ -6,7 +6,7 @@ import { microStxToStx } from '@app/common/stacks-utils';
 import { removeCommas } from '@app/common/token-utils';
 import { TransactionFormValues } from '@app/common/transactions/transaction-utils';
 import { useSelectedAsset } from '@app/pages/send-tokens/hooks/use-selected-asset';
-import { useCurrentAccountAvailableStxBalance } from '@app/store/accounts/account.hooks';
+import { useCurrentAccountAvailableStxBalance } from '@app/query/balance/balance.hooks';
 import { useCurrentAccountMempoolTransactionsBalance } from '@app/query/mempool/mempool.hooks';
 
 export function useSendAmountFieldActions({
