--- src/app/query/account/account.hooks.ts
@@ -1,6 +1,0 @@
-import { useGetAccountInfoQuery } from './account.query';
-
-export function useAccountInfo(address: string) {
-  const { data: accountInfo } = useGetAccountInfoQuery(address);
-  return accountInfo;
-}
