+++ src/app/store/accounts/account.hooks.ts
@@ -1,5 +1,5 @@
 import { useAtom } from 'jotai';
-import { useAtomValue, useUpdateAtom } from 'jotai/utils';
+import { useAtomValue } from 'jotai/utils';
 
 import {
   addressNetworkVersionState,
@@ -7,23 +7,15 @@ import {
 } from '@app/store/transactions/transaction';
 import {
   accountsWithAddressState,
-  currentAccountAvailableAnchoredStxBalanceState,
-  currentAccountBalancesUnanchoredState,
   currentAccountConfirmedTransactionsState,
-  currentAccountInfoState,
   currentAccountState,
   currentAccountStxAddressState,
   hasSwitchedAccountsState,
   hasCreatedAccountState,
-  refreshAccountDataState,
   transactionAccountIndexState,
 } from '@app/store/accounts/accounts';
 
 import { currentAccountIndexState } from '../wallet/wallet';
-
-export function useCurrentAccountAvailableStxBalance() {
-  return useAtomValue(currentAccountAvailableAnchoredStxBalanceState);
-}
 
 export function useAccountConfirmedTransactions() {
   return useAtomValue(currentAccountConfirmedTransactionsState);
@@ -63,16 +55,4 @@ export function useHasSwitchedAccounts() {
 
 export function useHasCreatedAccount() {
   return useAtom(hasCreatedAccountState);
-}
-
-export function useRefreshAccountData() {
-  return useUpdateAtom(refreshAccountDataState);
-}
-
-export function useSetAccountBalancesUnanchoredState() {
-  return useUpdateAtom(currentAccountBalancesUnanchoredState);
-}
-
-export function useSetAccountInfo() {
-  return useUpdateAtom(currentAccountInfoState);
 }
