--- src/app/store/common/common.hooks.ts
@@ -1,3 +1,0 @@
-import { queryClient } from 'jotai-query-toolkit';
-
-export const jotaiWrappedReactQueryQueryClient = queryClient;
