--- src/app/store/devtool/devtool.hooks.ts
@@ -1,6 +1,0 @@
-import { useAtomValue } from 'jotai/utils';
-import { queryClientAtom } from 'jotai-query-toolkit';
-
-export function useClientQuery() {
-  return useAtomValue(queryClientAtom);
-}
